Move out sites mapping into a separate class/file
and expose PSP/ELM importers
