+++ website/imports/sites/__init__.py
@@ -1,8 +1,13 @@
 from .hprd import HPRDImporter
 from .uniprot import GlycosylationUniprotImporter, OthersUniprotImporter
+from .psp import PhosphoSitePlusImporter
+from .elm import PhosphoELMImporter
+
 
 __all__ = [
     HPRDImporter,
     GlycosylationUniprotImporter,
-    OthersUniprotImporter
+    OthersUniprotImporter,
+    PhosphoSitePlusImporter,
+    PhosphoELMImporter,
 ]
