+++ website/imports/sites/site_importer.py
@@ -11,6 +11,7 @@ from database import get_or_create, create_key_model_dict
 from imports import Importer, protein_data as importers
 # those should be moved somewhere else
 from imports.protein_data import get_preferred_gene_isoform
+from imports.sites.site_mapper import SiteMapper
 from models import KinaseGroup, Kinase, Protein, Site, SiteType, BioModel, SiteSource, Gene
 
 
@@ -42,52 +43,6 @@ def get_or_create_kinases(chosen_kinases_names, known_kinases, known_kinase_grou
             kinases.add(known_kinases[name])
 
     return kinases, groups
-
-
-def find_all(longer_string: str, sub_string: str):
-    """Returns positions of all overlapping matches"""
-    position = -1
-    matches = []
-
-    while True:
-        position = longer_string.find(sub_string, position + 1)
-
-        if position == -1:
-            return matches
-
-        matches.append(position)
-
-
-class OneBasedPosition(int):
-    pass
-
-
-def map_site_to_isoform(site, isoform: Protein) -> List[OneBasedPosition]:
-    """Finds all occurrences of a site (by exact sequence match)
-    in provided sequence of an alternative isoform.
-
-    Original position of the site is used to highlight "suspicious" cases,
-    in which the matched site is far away (>50% of isoform length) from
-    the one of the original site. This is based on premise that most of
-    alternative isoform does not differ by more than 50% (TODO: prove this)
-
-    Returned positions are 1-based
-    """
-    matches = [
-        m + 1 + site.left_sequence_offset
-        for m in find_all(isoform.sequence, site.sequence)
-    ]
-
-    if len(matches) > 1:
-        warn(f'More than one match for: {site}')
-    if any(abs(position - site.position) > len(isoform.sequence) / 2 for position in matches):
-        warn(
-            f'This site {site} was found on {matches} positions, '
-            f'and some are quite far away from the position in '
-            f'original isoform: {site.position}.'
-        )
-
-    return matches
 
 
 class SiteImporter(Importer):
@@ -212,8 +167,10 @@ class SiteImporter(Importer):
         if len(sites) == 0:
             return sites
 
+        mapper = SiteMapper(self.proteins, self.repr_site)
+
         # sites loaded so far were explicitly defined in data files
-        mapped_sites = self._map_sites_by_sequence(sites)
+        mapped_sites = mapper.map_sites_by_sequence(sites)
 
         # from now, only sites which really appear in isoform sequences
         # in our database will be considered
@@ -222,84 +179,6 @@ class SiteImporter(Importer):
         mapped_sites.drop(columns=['sequence', 'left_sequence_offset'], inplace=True, errors='ignore')
 
         return mapped_sites
-
-    def _map_sites_by_sequence(self, sites: DataFrame) -> DataFrame:
-        """Given a site with an isoform it should occur in,
-        verify if the site really appears on the given position
-        in this isoform and find where in all other isoforms
-        given site appears (by exact match of +/-7 sequence span).
-
-        If a site does not appear on declared position in the
-        original isoform, emit a warning and try to find the
-        correct position (to overcome a potential sequence shift
-        which might be a result of different sequence versions).
-
-        Args:
-            sites: data frame with sites, having (at least) following columns:
-                   'sequence', 'position', 'refseq', 'residue', 'sequence_left_offset'
-
-        Returns:
-            Data frame of sites mapped to isoforms in database,
-            including the sites in isoforms provided on input,
-            if those has been confirmed or adjusted.
-        """
-        print('Mapping sites to isoforms')
-
-        mapped_sites = []
-        already_warned = set()
-
-        for site in tqdm(sites.itertuples(index=False), total=len(sites)):
-
-            site_id = self.repr_site(site)
-
-            if site.refseq not in self.proteins:
-                if site.refseq not in already_warned:
-                    warn(f'No protein with {site.refseq} for {site_id}')
-                    already_warned.add(site.refseq)
-                # TODO: fallback to gene directly (if there is site.gene, try it)
-                continue
-
-            protein = self.proteins[site.refseq]
-
-            gene = protein.gene
-
-            if gene and gene.isoforms:
-                isoforms_to_map = {self.proteins[isoform.refseq] for isoform in gene.isoforms}
-            else:
-                isoforms_to_map = {protein}
-
-            positions = {}
-
-            # find matches
-            for isoform in isoforms_to_map:
-                positions[isoform] = map_site_to_isoform(site, isoform)
-
-            original_isoform_matches = positions[protein]
-
-            if not original_isoform_matches:
-                warn(f'The site: {site_id} was not found in {protein.refseq}, '
-                     f'though it should appear in this isoform according to provided sites data.')
-            elif all(match_pos != site.position for match_pos in original_isoform_matches):
-                warn(f'The site: {site_id} does not appear on the exact given position in '
-                     f'{protein.refseq} isoform, though it was re-mapped to: {original_isoform_matches}.')
-
-            # create rows with sites
-            for isoform, matched_positions in positions.items():
-
-                if not matched_positions:
-                    continue
-
-                for position in matched_positions:
-
-                    # _replace() returns new namedtuple with replaced values;
-                    # it is not protected but hidden (to allow 'replace' field)
-                    new_site = site._replace(
-                        refseq=isoform.refseq,
-                        position=position
-                    )
-                    mapped_sites.append(new_site)
-
-        return DataFrame(mapped_sites)
 
     def add_site(self, refseq, position: int, residue, mod_type, pubmed_ids=None, kinases=None):
